fix: remove unused imports and update documentation for inference options
